+++ apps/desktop/package.json
@@ -22,7 +22,6 @@
 		"@sentry/vite-plugin": "^2.7.0",
 		"@tanstack/react-query": "^4.24.4",
 		"@tauri-apps/api": "1.3.0",
-		"comlink": "^4.4.1",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-router-dom": "6.9.0"
