+++ apps/desktop/src/vite-env.d.ts
@@ -1,5 +1,4 @@
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-comlink/client" />
 
 declare interface ImportMetaEnv {
 	VITE_OS: string;
