+++ interface/app/$libraryId/Explorer/FilePath/Thumb.tsx
@@ -2,6 +2,7 @@ import { getIcon, iconNames } from '@sd/assets/util';
 import clsx from 'clsx';
 import {
 	memo,
+	useCallback,
 	useEffect,
 	useLayoutEffect,
 	useMemo,
@@ -44,6 +45,7 @@ export interface ThumbProps {
 	className?: string;
 	frameClassName?: string;
 	childClassName?: string | ((type: ThumbType | `${ThumbType}`) => string | undefined);
+	isSidebarPreview?: boolean;
 }
 
 export const FileThumb = memo((props: ThumbProps) => {
@@ -67,16 +69,16 @@ export const FileThumb = memo((props: ThumbProps) => {
 		isDark ? classes.checkers : classes.checkersLight
 	);
 
-	const onLoad = () => setLoaded(true);
+	const onLoad = useCallback(() => setLoaded(true), []);
 
-	const onError = () => {
+	const onError = useCallback(() => {
 		setLoaded(false);
 		setThumbType((prevThumbType) =>
 			prevThumbType === ThumbType.Original && itemData.hasLocalThumbnail
 				? ThumbType.Thumbnail
 				: ThumbType.Icon
 		);
-	};
+	}, [itemData.hasLocalThumbnail]);
 
 	// useLayoutEffect is required to ensure the thumbType is always updated before the onError listener can execute,
 	// thus avoiding improper thumb types changes
@@ -197,6 +199,7 @@ export const FileThumb = memo((props: ThumbProps) => {
 												itemData.extension) ||
 											''
 										}
+										isSidebarPreview={props.isSidebarPreview}
 									/>
 								);
