+++ interface/components/TextViewer/prism.ts
@@ -2,6 +2,12 @@
 
 // WARNING: Import order matters
 
+window.Prism = window.Prism || {};
+Prism.manual = true;
+
+import "prismjs";
+import './prism.css';
+
 // Languages
 // Do not include default ones: markup, html, xml, svg, mathml, ssml, atom, rss, css, clike, javascript, js
 import 'prismjs/components/prism-applescript.js';
@@ -52,3 +58,30 @@ import 'prismjs/components/prism-typoscript.js';
 import 'prismjs/components/prism-vala.js';
 import 'prismjs/components/prism-yaml.js';
 import 'prismjs/components/prism-zig.js';
+
+// Mapping between extensions and prismjs language identifier
+// Only for those that are not already internally resolved by prismjs
+// https://prismjs.com/#supported-languages
+export const languageMapping = Object.entries({
+	applescript: ['scpt', 'scptd'],
+	// This is not entirely correct, but better than nothing:
+	// https://github.com/PrismJS/prism/issues/3656
+	// https://github.com/PrismJS/prism/issues/3660
+	sh: ['zsh', 'fish'],
+	c: ['h'],
+	cpp: ['hpp'],
+	js: ['mjs'],
+	crystal: ['cr'],
+	cs: ['csx'],
+	makefile: ['make'],
+	nim: ['nims'],
+	objc: ['m', 'mm'],
+	ocaml: ['ml', 'mli', 'mll', 'mly'],
+	perl: ['pl'],
+	php: ['php', 'php1', 'php2', 'php3', 'php4', 'php5', 'php6', 'phps', 'phpt', 'phtml'],
+	powershell: ['ps1', 'psd1', 'psm1'],
+	rust: ['rs']
+}).reduce<Map<string, string>>((mapping, [id, exts]) => {
+	for (const ext of exts) mapping.set(ext, id);
+	return mapping;
+}, new Map());
