+++ packages/config/vite/index.ts
@@ -1,6 +1,5 @@
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
-import { comlink } from 'vite-plugin-comlink';
 import { createHtmlPlugin } from 'vite-plugin-html';
 import svg from 'vite-plugin-svgr';
 import tsconfigPaths from 'vite-tsconfig-paths';
@@ -12,8 +11,7 @@ export default defineConfig({
 		svg({ svgrOptions: { icon: true } }),
 		createHtmlPlugin({
 			minify: true
-		}),
-		comlink()
+		})
 	],
 	css: {
 		modules: {
@@ -25,8 +23,5 @@ export default defineConfig({
 		sourcemap: true,
 		outDir: '../dist',
 		assetsDir: '.'
-	},
-	worker: {
-		plugins: [comlink()]
 	}
 });
